mkvmerge: don't re-derive AC3 framing for already-framed source containers
Fixes #1104 as re-deriving framing required caching packets and delaying
delivering a frame until it was clear there was no extension frame
afterwards. Source containers that cannot tell the end of the file for
each individual track (but only for the whole file) like Matroska would
result in the framer to have that last packet buffered until the end of
the file has been reached at which point a file split would have already
occurred – and the buffered packet would have belonged to an output file
now long since finished.