Fix crash if called from deleted dir
From a Linux terminal, whose current dir is a dir which has been deleted
in the meanwhile, a call of e.g. "mkvmerge -V" leads to a crash.

This is due to the bfs::absolute call here, which itself - according to
the regarding Boost doc - here results in a call to bfs::current_path
(left bottom case from the table). This call results in a crash as the
dir has already been deleted.

The used "/proc/self/exe" should (will?) always refer to a path already
being absolute. Therefore bfs::absolute now is only used if the path is
not yet absolute, in order to avoid such a crash.