GUI: Windows: store settings in INI files in user's data folder if installed
Includes migrating existing settings from the registry.

The background is that writing large amounts of data (as is done if the
queue's big) is kind of slow. This is in preparation for more changes,
e.g. storing each queue job in its own file instead of all queue jobs in
a single INI file.