merge: implement callback before adding packet to cluster
This lays the groundwork for implementing #1420. When the packetizer
notices garbage it doesn't know anything about the timecode the next
valid packet will end up with. That is affected by additional factors
like delays or even splitting files.

In effect the actual timecode is only known right before the packet is
added to a cluster. Therefore outputting warnings for the user that
include correct and human-understandable timecodes is only possible at
that moment as well.