mkvmerge: MP4: derive basic MP3/AC3 audio parameters from content
There are quite a lot of MP4 files which contain the wrong number of
channels in the track headers, especially for AC3 tracks (e.g. 2 instead
of 6 in one of my sample files). This change makes mkvmerge decode an
AC3 header from the content and uses that header's value instead of the
values stored in the track headers.