GUI: implement default track language by type
Before there was a single setting »default track language« that was
applied to all tracks regardless of their type.

This has been split up into separate settings for audio, video and
subtitle tracks. The rationale is that you often have a default language
for audio tracks but a different one for subtitles. In addition video
usually don't have a language at all for which zxx (no linguistic
content) might be appropriate as a default.

Implements #1338.