GUI: use regexes for chapter/tag/segmentinfo recognition when dropping files
Using the full-blown parsers for the supported file formats is very slow
as they read the whole file before doing anything with it. This is
obviously bad when input files are usually hundreds of MB big or bigger.

Instead only read the first 1 KB of data and do simple ASCII regex
matches against those bytes.