GUI: merge: don't leave empty tab when dropping files
If the current tab is empty then the mode »create new mux settings for
each dropped files« will put the first file into the current
tab. Otherwise an empty tab would be left.

Additionally when dropping a single file onto an otherwise empty tab the
GUI will always add that file to the tab. It won't ask the user even if
the drop mode is set to always ask, and it won't create new mux settings
even if the mode is set to do that. This should be more intuitive to
use.

Fixes #1446.