GUI: properly construct a file:// URL when opening the output folder
The GUI tells Qt to open a directory like
e.g. "file:////home/mosu/…". Two slashes are needed for the protocol,
one for the root folder, and the fourth one is superfluous.

Now file managers like Windows Explorer or krusader simply open a
directory name called "//home/mosu/…" – that works. But others like
Dolphin try to be smart, see that the file name starts with "//…" and
think the user wants to visit a share called "mosu" on a Samba/Windows
server called "home". Then they prepend the whole URL with "smb://".

On Windows the third slash is actually needed when letting Qt parse such
a URL. So instead of using #ifdefs create an empty URL and set the
scheme and the path components explicitly and don't let Qt parse a
stringified URL.

Fixes #1479.