GUI: jobs: assume jobType == MuxJob if not set in job file
Older versions of MKVToolNix didn't write a jobType attribute into a job
file. So newer versions checking for validity of that field failed an
assertion and aborted. The original intention was to convert old
settings on the fly, and that assertion made it impossible.

This was changed in the following commit

39634e8 GUI: jobs: ignore job files with missing/unknown jobType field when loading the queue

Now those jobs were simply ignored.

A better way would have been to simply assume a value of "MuxJob" if the
attribute is missing or empty as MuxJobs are the only ones the GUI
supported back when no jobType attribute was written.

See also #1481.