Matroska reader: use file analyzer if no seek head found before first cluster
This allows mkvmerge to find level 1 elements (attachments, chapters,
track headers, segment info, tags) located at the end of the file if
there's no seek head at the start, for whatever reason (e.g. bugs in
mkvpropedit).

As such analysis takes time that grows linearly with the file size it
won't be used for all files. It can be forced to be used with "--debug
kax_reader_use_analyzer", though.

See #1513 for a use-case.