mm_io_c: open MODE_CREATE files for reading and writing on Windows, too
mkvmerge's output file is opened in MODE_CREATE. This mode is supposed
to create the file if it doesn't exist, to truncate it otherwise and to
open it for writing. The fix for #1498 requires reading from the file,
too, and therefore the file must be opened for reading, too.

On non-Windows systems MODE_CREATE was already translated into opening
for reading and writing.