mkvmerge: add container ID as "number" attribute in verbose/JSON identification
Several container formats provide an internal ID of some kind. mkvmerge
already output that ID for some of the container formats it supports
albeit under different property names:

- Matroska: "number"
- MPEG program streams: "stream_id" and "sub_stream_id", encoded as a
  hexadecimal string instead of integers
- MPEG transport streams: "ts_pid"

With this commit the support is extended to other container formats. The
MPEG program and transport stream reader modules output their IDs as the
property "number" (in addition to the older identifiers) in order to
have a consistent naming scheme which makes using the output easier for
other applications.

The changes are:

- QuickTime/MP4: the track ID from the 'tkhd' atom
- MPEG program stream: the sub-stream ID in the upper 32 bits and the
  stream ID in the lower 32 bits
- MPEG transport stream: the program ID
- Ogg/OGM: the stream's serial number field
- RealMedia: the track ID

Implements #1541.