GUI: watch jobs: re-add "clear output" in "more actions" menu
This regression was introduced in aeb0add.

The new menu is created during the tab's constructor. At that point it
called "isCurrentJobTab()" do determine whether or not to add the "clear
outputs" action. However, "isCurrentJobTab()" relied on the MainWindow's
currentJobTab variable being set, which at that point isn't yet as the
tab is still being constructed.

Solve this by passing a static parameter to the tabs that indicate
whether or not they're the "current job" tab.