GUI: attachments: change page storage from list to hash
Each page has an ID when it's stored in the model. This ID used to be
the position in the list. This was fine as long as the list wasn't
dynamic. However, with the new attachment editing features the user can
add and remove attachments in arbitrary order. This means that the
position inside the list changes but not the ID associated with the
page.

Therefore a hash is now used instead with a sequential number as the ID.