bit_cursor: allow setting bit position to exactly the end
For one it makes for easier-to-use interfaces such as
bc.set_bit_position(bc.get_bits…)) without having to check if the
position is at the end and the get_bits returns 0. It's also consistent
with containers in the standard C++ library.

In this particular case it fixes the AAC handling of program config
elements with an empty comment field that are located at the end of the
GA specific config where the aforementioned pattern of
set_bit_position(get_bits(…)) is used.

Fixes #1578.