mkvextract: fix chapter extraction
The old code was horribly broken in several ways:

- Deeper levels of atoms were somehow used, too.
- Whenever an atom contained multiple display entries then a line was
  output for each name/language pair instead of one line for the whole
  atom.
- The wrong number of entries was output whenever multiple names were
  used.

The new code only processes the top-most level of chapter atoms of each
edition. For each atom only one line is output, and the first name found
for that atom is output as the chapter's name.

Fixes the bug aspect of #1610.