kax_analayzer_c: handle last level 1 element having an unknown size
If this is the case then the size must be fixed to a known one before
further modification can take place. Otherwise placing an element behind
the one with the unknown size would make the new element be part of the
one with the unknown size.

Such an operation can only succeed if the unknown-sized level 1
element's size has been written with enough bytes for the new element
size's length.

Such files are produced by e.g. HandBrake which doesn't write cues or
meta seek elements at the end. Instead the last element is a cluster
element with a eight-byte long size field set to »unknown«.

Fixes #1601.