xtr_hevc_c: don't write VPS, SPS, PPS, SEI NALUs twice
Certain NALUs (VPS, SPS, PPS) are stored entirely in the HEVCC (codec
private data). Those are written to the file first during
extraction. However, they may also be stored in front of the first frame
leading to a duplication.

mkvextract now skips VPS, SPS and PPS NALUs located before the first
frame during extraction.

The situation with SEI NALUs is slightly different. Only certain types
of SEI NALUs are actually stored in the HEVCC. In order not to duplicate
them mkvextract will check whether or not the first frame contains
SEIs. If it does then the SEIs stored in the HEVCC won't be written to
the output file, but the SEIs from the first frame are. Otherwise (no
SEIs in the first frame) the SEIs from the HEVCC are written.

See #1076 and #1621.