teletext subtitles: collapse multiple entries with same content into single entry
The sample file in question contains many teletext packets that show the
same content. The old code would convert those to separate Matroska
frames.

Additionally each frame's duration was calculated from the wrong
timestamp. A teletext page can be drawn over the space of several
packets with different timestamps. Therefore the timestamp of the first
packet rendering a line of a teletext page must be used as the start
timestamp, and the timestamp of the first packet rendering a line of the
following teletext page as its end.

The old code was using a wrong packet for the end, though: the timestamp
of packet containing the end-of-page block.

Fixes #1623.