mkvmerge: update chapter size before deciding to overwrite chapter void
Without a call to UpdateSize() the actual size of the chapters to be
written would not be known. Therefore the decision whether or not to
overwrite the void placeholder reserved for chapters would be made based
on wrong data. In effect mkvmerge was always trying to overwrite the
void which ultimately failed because it was actually calling
UpdateSize() and saw that the new elements wouldn't fit.

This affected chapters generated with --generate-chapters but not
chapters added through other means (e.g. --chapters or when reading them
from other source files). For the latter other code paths were already
calling UpdateSize() in appropriate places.