subtitle packetizer: assume system character set if none is given
This is a regression introduced in
8ca1160de08bd281cf707cda75417dea1d80ec51 which was supposed to fix
--sub-charset having an effect on files with a byte-order
mark (BOM).

Problem is that the supposed fix also made the subtitle packetizer only
recode if a character set was explicitly given even for files which do
not contain a BOM. Instead it should recode from the system's character
set in such a case. This is done if charset_converter_c::init() is
called with an empty string. Unfortunately commit 8ca1160 changed the
packetizer not to call charset_converter_c::init() if the given
character set is empty.

Thee semantics are:

- "recode" is not set if the source is known to be in UTF-8 already,
  e.g. if the track is read from an SRT file with a BOM or if it's read
  from a Matroska file.
- "is_utf8" is set if the source is known to be in UTF-8 for other
  reasons.

Therefore if recode is set and is_utf8 isn't then the conversion must
take place no matter if --sub-charset is given or not. If it is
given (m_ti.m_sub_charset is not empty) then use that;
otherwise (m__ti.m_sub_charset is empty) use the system's character
set.

Fixes #1639.