Mac OS build: link with -headerpad_max_install_names
The executable headers on Mac OS contain paths to shared libraries. When
building the disk image these absolute paths must be changed into ones
relative to the executable's location so that the end user doesn't have
to install those libraries (Qt in this case) system-wide.

Unfortunately there may not be enough room in the headers for this
substitution. The additional linker flag "-headerpad_max_install_names"
forces the linker to reserve enough space.