QuickTime/MP4 reader: JSON ident: report audio sampling frequency as an integer
My JSON schema says it should be an integer, so make it one and not a
floating point number.

All other input modules already either use integers for storing the
value or cast the floating point variables to integers before adding
them to the JSON output.

This will only required required with an updated version of nlohmann's
JSON library as the current library implicitly casts the float to an
int. The updated version won't do that anymore and preserve the type.