chapter generation: handle failing void replacement by writing chapters at end
The overwrite function already takes care of updating the element's size
properly. Therefore it doesn't have to be done by mkvmerge before
calling the function, especially if it's called wrong:

- our manual call doesn't include elements with default value in the
  calculation
- the replacement request does

This led to the situation where mkvmerge thought that the new chapters
would fit into the void while the replacement function disagreed causing
the chapters not the be written at all.

Fixes #1693.