GUI: temporarily allow colon as first character in output file name
That way changing the drive letter by removing and typing a new one
becomes possible again. When starting to mux/adding the job to the queue
the output file name is checked again, and if the drive letter is
missing then the GUI will show an appropriate error message.

Fixes #1692.