MPEG PS, TS: make probe size depend on source file size
Instead of using a fixed maximum both readers now use a limit that's
based on a percentage of the source file's size. There's still a
hard-coded lower limit of 5 MB, though.

This is in preparation of allowing the user to change the probe range by
specifying a percentage of the source file size. With that the user can
solve situations where certain tracks start very late in the file
without impacting performance for every user (e.g. by making mkvmerge
always scan the whole file).

See #1734.