GUI: merge: copy m_probeRangePercentage member when cloning a SourceFile
Not copying the field will leave the member in the target uninitialized
and therefore more or less random. This can lead to interesting values
being present in the field.

In addition with the wrong format specifier for converting a double into
a string this leads to errors from mkvmerge about the probe range being
invalid if, due to the actual data in the uninitialized field, the
stringified version is written with scientific notation.

Part of the fix for #1741.