GUI: jobs: import dropped .mtxcfg files into the job queue
When the user drags & drops a valid .mtxcfg file (either a full job file
or one containing only the merge settings without the job attributes)
the GUI will not import the content into the job queue as a new job. The
job's properties will be reset (no output/errors/warnings, date fields
are cleared/reset to the current date), and its initial state will be
"pending manual start".

Rest of the implementation of #1714.