merge: update video packetizer reference when appending
The global video packetizer reference is used in several decisions. One
of them is the decision when to start a new cluster. For determining
whether a frame is a video frame that global video packetizer reference
is compared to the packet's source packetizer. If they match (and if
it's a key frame) then starting a new cluster is now considered.

However, each appended file has its own packetizer instance. Therefore
the comparison above fails starting with the first appended file and all
following files.

Fixes #1757.