merge: force-pull all packetizers from a fully-held file
In its core mkvmerge uses a pull model: it iterates over each packetizer
and asks it to produce a packet. The packetizer passes that request on
to the reader it belongs to. The reader will then read the next frame
for the track corresponding to the packetizer and pass that data back to
the packetizer.

However, several container formats (e.g. MPEG PS, Matroska, Ogg and
others) do not allow arbitrary access to the frames for each
track. Instead mkvmerge has to read them from front to back. In such a
case the reader ignores the packetizer and simply reads the next frame,
whichever packetizer that might belong to.

This is problematic with sparse tracks. For those the packetizer will
repeatedly call the reader's read() method, but the read() method will
only produce data for the other tracks in that file. That data is then
buffered in those tracks' packetizers.

In order to limit that memory usage the readers/packetizers know a state
known as "holding" — meaning a lot of data has been buffered and no data
for this packetizer has been found yet.

The actual problem occurred when all packetizers for a file were holding
and no packets were present for any of those packetizers:

- The packetizers run empty on data.
- The reader sees too much data had been buffered and enters "holding"
  state.
- The core only saw packetizers that don't have any data anymore and
  concluded that it was done.

This commit changes the behavior to force a pull in such a
situation (which the readers already provided for — it was just that the
code never opted to force the pull). This will allow the reader to
exceed its allotted memory usage and deliver more data.