AVC, HEVC readers: fix requiring detected width/height to be positive
Under certain conditions the width/height of hadn't been known in the
HEVC when the HEVCC was considered to be ready. The width/height were
only fixed by later SPS instances in the bitstream.

In that case the HEVC reader was only reading until the the parser
signaled that the HEVCC was ready. The following check for width &
height being positive failed, causing detection to fail.

The change is to let the HEVC parser signal readiness if and only if
both the HEVCC is ready and width & height are positive.

The same applies to the AVC reader.

This fixes the regression introduced in
508c925eef8c30c228a24a474965544ff0ffb185.