GUI: make tool names consistent
Rename "Merge" to "Merge tool" in both the main window and the
preferences. Makes translations easier, and I'm talking about the "merge
tool" everywhere else already.

Rename "edit chapters" and "edit headers" to "chapter editor" and
"header editor" in the main window for similar reasons. In the
preferences the latter terms had already in use anyway.