WAV reader: require 5 consecutive DTS headers during byte swap/14-in-16 detection
The original code tried all four possible combinations of byte swapping
yes/no and 14-in-16 yes/no and aborted as soon as a single DTS header
was found. This can be a false positive, though, and with one test file
provided by a user this has indeed been the case.

The new code does the same detection but requires five consecutive DTS
headers to be found before the detection is deemed to be successful.