h.264: include all SPS/PPS NALUs before key frames after changed SPS/PPS content
In some h.264 bitstreams the SPS and/or PPS settings change, meaning the
bitstream contains SPS/PPS NALUs with an ID that was encountered before,
but with different content. The specs say that the newly encountered
SPS/PPS settings are to be used from that point on.

So far this wasn't properly handled by mkvmerge. It would only include
the changed SPS/PPS NALUs at the point where the change
occurred. However, an application seeking to a point after such a change
would have no knowledge of the change and use the SPS/PPS from the AVCC
stored in CodecPrivate. It would therefore try to decode the pictures
with the wrong settings leading to image corruption and worse.

With this change mkvmerge will include all currently active SPS and PPS
NALUs in front of every key frame (every frame marked as "key" in
Matroska, that is) once it encounters at least one SPS or PPS NALU with
changed content.

Fixes #1711.