generic_packetizer_c: add callback for after timestamping a packet
The intention is to enable packetizers to do something with the modified
timestamp and/or duration. For example, in VobSubs the duration is also
stored in the bitstream and should be updated. However, arguments such
as --sync and --timecodes are handled by the common packetizer
functionality, not by format-specific packetizer code. Hence the
callback.

This is in preparation of fixing #1771.