mkvmerge: handle end-of-input in "force pulling" function, too
Otherwise a situation may arise where a file still in a non-finished
state going into the force pull function and coming out of it in a
finished state. However, mkvmerge has to do some work on such a
transition:

• housekeeping (decrease number of unfinished packetizers for later
  decisions whether or not a file has been read)
• establishing deferred connections when appending

In such a case mkvmerge entered an endless loop.

This was a regression introduced with commit
e06d80f70b564ef130de8eee53f551b80d91da31 as a fix to #1774.