GUI: don't use lock file for instance communication socket
The lock file has proven to be a source of problems time and again on
Windows, see e.g. #1805. In most instances a stale lock file was still
present though the process it referred to wasn't running anymore. Qt is
supposed to clean up such stale lock files automatically, but that
didn't work reliably. The result from the user's POV was that the GUI
simply didn't start as neither the prior instance was still running nor
was the new one showing its UI.

Therefore we just skip trying to lock that file. Instead the new
instance will simply try to connect to the communication socket. If that
connection succeeds, it can be reasonably sure another instance is
running. If it fails, though, the GUI simply removes any lingering
socket file and tries to set it up again. Even if the removal &
subsequent setup fails, the GUI should still come up as a second
instance.