Update de.po
I have never ever heard the term "verschränkt" as a translation of interlaced. And I'm probably not alone: Searching on forum.gleitz.info for "verschränkt" found three topics (two of which actually use it in the sense of interlaced (but I'd say that neither of these topics uses "verschränkt" as German for "interlaced")), but 500 for interlaced (that's probably the maximum number of search results one can get, so there are probably more than 500 topics). So I propose to change it to "Zeilensprungverfahren" instead.