mkvmerge: assign track numbers for tracks in --track-order first
When the user specifies tracks with "--track-order a,b,c…", she expects
track a to be assigned track number 1, b to be assigned track number 2
and 3 for track c.

The old code did not guarantee this. Instead it might assign track ID 1
to a track not listed in "--track-order".

Fixes the second part of #1832.