Update README.md
In order to make the text more beautiful and easier to read/understand for readers, I have added the following changes:

1. Added the word **via** before the product names in brackets → (**via** mkvinfo), (**via** mkvextract), (**via** mkvmerge). The word **via** clearly indicates that these tools are meant to be used for something, and they are not lonely.

2. Added many missing commas, inter alia:

a) before the word **then**. In English we (almost) always add a comma before **then** if it is located in the middle of a sentence. The word **then** is often seen in English conditional sentences, but in 99% cases it's optional, so we can use a subject directly (see also: http://images.slideplayer.com/28/9327486/slides/slide_3.jpg)

For more information about **then** and English conditional sentences, you can visit the following websites: 

• https://en.wikipedia.org/wiki/English_conditional_sentences

• https://en.wikipedia.org/wiki/Conditional_sentence 

b) before a subject (e.g. **I**, **you**) if it is located in the middle of a sentence. For example:  **If you want to install it yourself, I suggest you use..........** In such cases we always add a comma in order to emphasize the difference between two sentences (two subjects); in the first one: **you** and in the second one: **I**. Thanks to the comma we can easily distinguish them when reading the text.

c) before an (ad)verb if a sentence starts with **If**. For example: **If at all possible, please include.......**. In this case I have added the comma in order to separate an adjective (**possible**) and an (ad)verb (**please**). And also before the adverb **there** which is located in the middle of a sentence and before a noun (**development**). For example: **As both Matroska and MKVToolNix are under heavy development, there might be features......**. 

3. Removed one unnecessary comma which was located before **if** (when the **if-clause** is at the end) → **e.g. include the command line, if you use Windows or Linux etc.pp.** In English we don't generally add a comma if the word **if** is located in the middle of a sentence. See here for rationale: http://images.slideplayer.com/11/3193271/slides/slide_4.jpg and also above for the English Wikipedia (conditional sentences).