Update README.Windows.md
I have added the following changes into the text:

1. Added many missing commas, inter alia:

a) before the word **then**. In English we (almost) always add a comma before **then** if it is located in the middle of a sentence. The word **then** is often seen in English conditional sentences, but in 99% cases it's optional, so we can use a subject directly (see also: http://images.slideplayer.com/28/9327486/slides/slide_3.jpg)

For more information about **then** and English conditional sentences, you can visit the following websites:

• https://en.wikipedia.org/wiki/English_conditional_sentences

• https://en.wikipedia.org/wiki/Conditional_sentence

b) before a subject (e.g. **I**, **you**, **the author**) if it is located in the middle of a sentence. In such cases we always add a comma in order to emphasize the difference between two sentences (two subjects). Thanks to the comma we can easily distinguish them when reading the text.