GUI: move update check to its own thread
Normally QNetworkAccessManager is supposed to provide an asynchronous
interface to network requests. Unfortunately, the very first call to
QNetworkAccessManager::get() for each URL takes quite a long time,
e.g. more than two seconds on my Windows installation. This is even
before the request is actually started. As this happens on the main
thread it blocks the UI.

In MKVToolNix the update check is started during the GUI's startup
process. Any kind of blockage of the main thread is very noticeable and
annoying.

This seems to be due to the following but in Qt:
https://bugreports.qt.io/browse/QTBUG-10106