Rakefile: enable parallel builds with rake 10.1.0 or newer
rake 10.1.0, which is part of Ruby v2.1.x, can do parallel builds, too,
and supports the same option -j that drake does. And it even defaults to
"number of cores + 1".

However, the tasks that can be executed in parallel have to be set up
with "multitask" instead of "task". There is an option to enable
parallel processing for all tasks defined as "tasks", though, and the
commit enables this option.