mkvmerge cleanup: discard write buffer content in error situations
This prevents the error message "not enough space on disk" being shown
twice.

Whenever a write fails, an exception is throw, and an appropriate error
message is shown. This is the first time.

Next mkvmerge goes into its cleanup routine. There it closes the output
file. The output file uses a write buffer implementation. Before closing
the file any outstanding buffered content is written to the disk. The
disk is still full, though, triggering another exception, and another
error message is output.

The workaround is to discard any buffered content still remaining at
cleanup time. This is safe as the output file is closed manually normal
program flow. Therefore no buffered content is lost in a successful run.

Fixes #1850.