GUI: don't lock cache dir for whole duration of cleanup operation
With a large number of files cleaning the cache can take quite some
time. During that time file identification won't work as it tries to
acquire a lock that's already held by the cleanup process.

With this change the cleanup process will release the lock after having
processed each file allowing the identification process to obtain the
lock temporarily.

Fixes #1860.